Add additional helper scripts for database set-up:
Two remaining issues in making it easy to set up the database
for the system are:

* Only having MySQL SQL files.
* Table code along with starter data in one file.

I've added two scripts here to help:

* `bin/generate_sql.sh`: create SQLite SQL files on the fly from
  the schema.
* `db_patches/set_up_data.sql`: a file that simply contains the
  `INSERT` statements.

These are offered mainly as a suggestion and should be taken
with a grain of salt. Is this the best way to handle this? Should
these be generalized?