BLOGSPERL-84 Match assets with existing blog posts
The existing blog posts use 'http://blogs.perl.org' as the start of the asset
URLs, rather than local references. This patch edits the blog post text to
change the asset URLs to point to the local box. It's probably easier to just
use relative URLs, but that might cause other complications.