BLOGSPERL-234 Avatars are not correctly loaded at search/posts  : Solved & Tested.
The JSON objects which are sent when searching contain only the dark
avatar. So, the situation when the theme was already set to light but on
seach avatars are dark appeared.
By setting the default to null, at search the avatars will be set
according to the theme slider -  functionaly which was already made by
Dan.